improve annotation query interface
previously, the only enforcement that annotation searches must include
tasks was at submit-time validation, and programmers were required
to read the documentation or suffer the consequences.

to fix this, introduce a new fluent class for constructing annotation-
related query terms, where the builder for this class requires a
task list, and change the document searcher to accept an instance
of this new class as its annotation search.

when used, this will look like:

```
  import static com.idibon.api.model.DocumentAnnotationQuery.forTasks;
  import static com.idibon.api.model.Annotation.Provenance.*;

  myCollection.documents().annotated(forTasks("MYTASK").by(Human))
```