add significant feature reporting to SDK
get rid of the Predictable interface, since it made adding common
query parameters for predictions (such as significant feature
reporting) more cumbersome. also, as implemented there was no
verification that the document was in the same collection as the
task, so cross-predictions would have failed.

fix both problems by switching over to DocumentContent as the
underlying interface for predictions (and uploads), and adding
special handling for static predictions by checking for Document
instances directly